Fix compile error on lkgr Win x64 build.
R=calamity@chromium.org

Review URL: https://codereview.chromium.org/24856003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@225669 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1d1e24eff98ad57a5fe7e8a3a6bdabc4164afd24

+++ content/plugin/webplugin_delegate_stub.cc
@@ -435,7 +435,7 @@ void WebPluginDelegateStub::OnFetchURL(
                       params.first_party_for_cookies,
                       params.method,
                       data,
-                      params.post_data.size(),
+                      static_cast<unsigned int>(params.post_data.size()),
                       params.referrer,
                       params.notify_redirect,
                       params.is_plugin_src_load,
